#104: Flyttar krypter-metode til test, sia den ikkje brukast i prodkode
