+++ src/test/kotlin/no/roedt/ringesentralen/hypersys/AESUtilsTest.kt
@@ -6,6 +6,11 @@ import com.nhaarman.mockitokotlin2.whenever
 import no.roedt.ringesentralen.token.SecretFactory
 import org.junit.jupiter.api.Assertions
 import org.junit.jupiter.api.Test
+import java.security.SecureRandom
+import java.util.*
+import javax.crypto.Cipher
+import javax.crypto.SecretKey
+import javax.crypto.spec.IvParameterSpec
 
 internal class AESUtilUnitTest {
 
@@ -19,9 +24,19 @@ internal class AESUtilUnitTest {
 
         val aesUtil = AESUtil(secretFactory).also { it.setKey() }
 
-        val cipherText = aesUtil.encrypt(input)
+        val cipherText = encrypt(input = input, algorithm = aesUtil.algorithm, key = aesUtil.key)
         val plainText = aesUtil.decrypt(cipherText)
 
         Assertions.assertEquals(input, plainText)
+    }
+
+    private fun encrypt(input: String, algorithm: String, key: SecretKey): String {
+        val iv = IvParameterSpec(ByteArray(16).also { SecureRandom().nextBytes(it) })
+        val encoder = Base64.getEncoder()
+
+        return Cipher.getInstance(algorithm)
+            .also { it.init(Cipher.ENCRYPT_MODE, key, iv) }
+            .doFinal(input.encodeToByteArray())
+            .let { encoder.encodeToString(iv.iv)  +":" + encoder.encodeToString(it) }
     }
 }
